requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coremltools==3.3
ffmpeg-python==0.2.0
future==0.18.2
idna==2.9
linecache2==1.0.0
nose==1.3.7
numpy==1.18.2
numpy-stl==2.10.1
pandas==1.0.2
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
rectpack==0.2.1
six==1.14.0
soupsieve==1.9.5
traceback2==1.4.0
trimesh==3.6.5
unittest2==1.1.0
urllib3==1.25.8
